[tests] move LitNode to separate module
